+++ test/credentials/verification/vc_integrity_verifier_test.dart
@@ -5,6 +5,28 @@ import '../../fixtures/verifiable_credentials_data_fixtures.dart';
 
 void main() {
   group('VC Integrity Verifier', () {
+    test('Should pass for valid VC issues by other libs', () async {
+      final vcList = [
+        VerifiableCredentialDataFixtures
+            .credentialWithEcdsaSecp256k1Signature2019ByVault,
+        VerifiableCredentialDataFixtures
+            .credentialWithEcdsaRdfc2019ByDigitalBazaar,
+        VerifiableCredentialDataFixtures
+            .credentialWithEddsaRdfc2022ByDigitalBazaar
+      ];
+
+      for (final vc in vcList) {
+        final verifiableCredential = UniversalParser.parse(vc);
+
+        final verifier = VcIntegrityVerifier();
+        var result = await verifier.verify(verifiableCredential);
+
+        expect(result.isValid, true);
+        expect(result.errors, <String>[]);
+        expect(result.warnings, <String>[]);
+      }
+    });
+
     test('Should pass for valid VC from CWE', () async {
       var data = VerifiableCredentialDataFixtures.ldVcDm1ValidStringFromCwe;
       final verifiableCredential = UniversalParser.parse(data);
@@ -17,7 +39,7 @@ void main() {
       expect(result.warnings, <String>[]);
     });
 
-    test('should failed verification for jwt dm v1 for invalid signature',
+    test('Should failed verification for jwt dm v1 for invalid signature',
         () async {
       final verifier = VcIntegrityVerifier();
       var data =
@@ -31,7 +53,7 @@ void main() {
       expect(result.warnings, <String>[]);
     });
 
-    test('should pass verification for sdjwt', () async {
+    test('Should pass verification for sdjwt', () async {
       final verifier = VcIntegrityVerifier();
       var data = VerifiableCredentialDataFixtures.sdJwtWithValidSig;
       final verifiableCredential = UniversalParser.parse(data);